Storage fixes, improvements and tests

+++ src/Sitko.Core.Storage.S3/S3Storage.cs
@@ -1,5 +1,7 @@
 using System;
+using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Threading.Tasks;
 using Amazon;
 using Amazon.S3;
@@ -10,16 +12,15 @@ using Microsoft.Extensions.Logging;
 
 namespace Sitko.Core.Storage.S3
 {
-    public sealed class S3Storage<T> : Storage<T>, IDisposable where T : IS3StorageOptions
+    public sealed class S3Storage<T> : Storage<T> where T : IS3StorageOptions
     {
-        private readonly ILogger<S3Storage<T>> _logger;
         private readonly T _options;
         private readonly AmazonS3Client _client;
         private bool _disposed;
+        private readonly List<FileStream> _openedStreams = new List<FileStream>();
 
         public S3Storage(T options, ILogger<S3Storage<T>> logger) : base(options, logger)
         {
-            _logger = logger;
             _options = options;
 
             var config = new AmazonS3Config
@@ -33,22 +34,11 @@ namespace Sitko.Core.Storage.S3
             _client = new AmazonS3Client(_options.AccessKey, _options.SecretKey, config);
         }
 
-        public void Dispose()
-        {
-            if (_disposed)
-            {
-                return;
-            }
-
-            _disposed = true;
-            _client.Dispose();
-        }
-
         private async Task CreateBucketAsync(string bucketName)
         {
             try
             {
-                var bucketExists = await AmazonS3Util.DoesS3BucketExistV2Async(_client, bucketName);
+                var bucketExists = await IsBucketExists(bucketName);
                 if (!bucketExists)
                 {
                     var putBucketRequest = new PutBucketRequest {BucketName = bucketName, UseClientRegion = true};
@@ -58,9 +48,23 @@ namespace Sitko.Core.Storage.S3
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, ex.Message);
+                Logger.LogError(ex, ex.Message);
                 throw;
             }
+        }
+
+        private Task<bool> IsBucketExists(string bucketName)
+        {
+            return AmazonS3Util.DoesS3BucketExistV2Async(_client, bucketName);
+        }
+
+        private async Task<bool> IsObjectExists(string filePath)
+        {
+            var request = new ListObjectsRequest {BucketName = _options.Bucket, Prefix = filePath, MaxKeys = 1};
+
+            var response = await _client.ListObjectsAsync(request);
+
+            return response.S3Objects.Any();
         }
 
         protected override async Task<bool> DoSaveAsync(string path, Stream file)
@@ -75,23 +79,71 @@ namespace Sitko.Core.Storage.S3
             }
             catch (Exception e)
             {
-                _logger.LogError(e, e.Message);
+                Logger.LogError(e, e.Message);
                 throw;
             }
         }
 
         public override async Task<bool> DeleteFileAsync(string filePath)
         {
-            try
+            if (await IsBucketExists(_options.Bucket))
             {
-                await _client.DeleteObjectAsync(_options.Bucket, filePath);
-                return true;
+                if (await IsObjectExists(filePath))
+                {
+                    try
+                    {
+                        await _client.DeleteObjectAsync(_options.Bucket, filePath);
+                        return true;
+                    }
+                    catch (Exception e)
+                    {
+                        Logger.LogError(e, e.Message);
+                    }
+                }
             }
-            catch (Exception e)
+
+            return false;
+        }
+
+        public override async Task<Stream> DownloadFileAsync(StorageItem item)
+        {
+            var path = Path.GetTempFileName();
+            using (var utility = new TransferUtility(_client))
             {
-                _logger.LogError(e, e.Message);
-                throw;
+                await utility.DownloadAsync(path, _options.Bucket, item.FilePath);
             }
+
+            var stream = File.OpenRead(path);
+            _openedStreams.Add(stream);
+            return stream;
+        }
+
+        public override async Task DeleteAllAsync()
+        {
+            if (await IsBucketExists(_options.Bucket))
+            {
+                await AmazonS3Util.DeleteS3BucketWithObjectsAsync(_client, _options.Bucket);
+            }
+        }
+
+        private void ClearStreams()
+        {
+            if (_openedStreams.Any())
+            {
+                foreach (FileStream stream in _openedStreams)
+                {
+                    stream.Close();
+                }
+
+                _openedStreams.Clear();
+            }
+        }
+
+        public override ValueTask DisposeAsync()
+        {
+            _client.Dispose();
+            ClearStreams();
+            return base.DisposeAsync();
         }
     }
 }
